+++ android/app/build.gradle.kts
@@ -112,7 +112,7 @@ dependencies {
     implementation(libs.compose.material3)
     implementation(libs.compose.ui)
     implementation(libs.compose.ui.preview)
-    implementation("com.slack.circuit:circuit-foundation:0.19.1")
+    implementation(libs.circuit.foundation)
     implementation(platform(libs.firebase.bom))
     implementation(libs.firebase.analytics)
     implementation(libs.firebase.crashlytics)
