+++ desktop/app/src/main/kotlin/Main.kt
@@ -30,13 +30,12 @@ import com.github.scribejava.core.oauth.OAuth20Service
 import com.slack.circuit.foundation.Circuit
 import com.slack.circuit.foundation.CircuitCompositionLocals
 import com.slack.circuit.foundation.CircuitContent
-import ly.david.musicsearch.feature.search.internal.SearchScreen
+import ly.david.musicsearch.feature.search.SearchScreen
 import ly.david.musicsearch.shared.di.sharedModule
 import ly.david.musicsearch.strings.LocalStrings
 import ly.david.ui.core.theme.BaseTheme
 import ly.david.ui.core.theme.TextStyles
 import org.koin.core.context.startKoin
-import org.koin.java.KoinJavaComponent.inject
 
 fun main() = application {
     val windowState = rememberWindowState()
