+++ feature/details/build.gradle.kts
@@ -26,7 +26,7 @@ kotlin {
                 implementation(compose.ui)
                 implementation(compose.preview)
 
-                implementation("com.slack.circuit:circuit-foundation:0.19.1")
+                implementation(libs.circuit.foundation)
                 implementation(libs.koin.core)
                 implementation(libs.kotlinx.collections.immutable)
             }
