+++ feature/search/index.html
@@ -310,7 +310,7 @@ footer {
             <td>3</td>
           </tr>
           <tr>
-            <td>ly.david.musicsearch.feature.search.internal.SearchScreen</td>
+            <td>ly.david.musicsearch.feature.search.SearchScreen</td>
             <td>SearchScreen</td>
             <td>1</td>
             <td>1</td>
