+++ feature/search/search_release-composables.csv
@@ -4,7 +4,7 @@ ly.david.musicsearch.feature.search.internal.SearchHistoryListItem,SearchHistory
 ly.david.musicsearch.feature.search.internal.PreviewSearchHistoryListItem,PreviewSearchHistoryListItem,1,1,1,0,0,0,0,0,1,1,
 ly.david.musicsearch.feature.search.internal.SearchHistoryScreen,SearchHistoryScreen,1,1,1,0,0,0,1,0,3,5,
 ly.david.musicsearch.feature.search.internal.SearchResultsScreen,SearchResultsScreen,1,1,1,0,0,0,1,0,2,3,
-ly.david.musicsearch.feature.search.internal.SearchScreen,SearchScreen,1,1,1,0,0,0,1,0,2,12,
+ly.david.musicsearch.feature.search.SearchScreen,SearchScreen,1,1,1,0,0,0,1,0,2,12,
 ly.david.musicsearch.feature.search.preview.PreviewRecentSearchesHeader,PreviewRecentSearchesHeader,1,1,1,0,0,0,0,0,1,1,
 ly.david.musicsearch.feature.search.internal.RecentSearchesHeader,RecentSearchesHeader,1,1,1,0,0,0,0,0,1,2,
 ly.david.musicsearch.feature.search.internal.PreviewRecentSearchesHeader,PreviewRecentSearchesHeader,1,1,1,0,0,0,0,0,1,1,
