+++ shared/build.gradle.kts
@@ -28,7 +28,7 @@ kotlin {
                 implementation(projects.ui.common)
                 implementation(projects.ui.history)
                 implementation(libs.koin.core)
-                implementation("com.slack.circuit:circuit-foundation:0.19.1")
+                implementation(libs.circuit.foundation)
             }
         }
         val androidMain by getting {
