+++ shared/src/commonMain/kotlin/ly/david/musicsearch/shared/di/SharedCommonModule.kt
@@ -18,7 +18,7 @@ import ly.david.musicsearch.data.musicbrainz.di.musicBrainzDataModule
 import ly.david.musicsearch.data.repository.di.repositoryDataModule
 import ly.david.musicsearch.data.spotify.di.spotifyApiModule
 import ly.david.musicsearch.data.spotify.di.spotifyDataModule
-import ly.david.musicsearch.feature.search.di.searchFeatureModule
+import ly.david.musicsearch.feature.search.searchFeatureModule
 import ly.david.musicsearch.feature.stats.di.statsFeatureModule
 import ly.david.musicsearch.strings.di.stringsModule
 import ly.david.ui.common.commonUiModule
